refactor: rename cosmoz-resize-handle files to resize-handle
Aligns with resizable-view naming convention (no cosmoz- prefix in filenames).
The custom element name 'cosmoz-resize-handle' is unchanged.
